+++ prqlc/prql-compiler/src/codegen/types.rs
@@ -38,14 +38,24 @@ impl WriteSource for TyKind {
             Ident(ident) => ident.write(opt),
             Primitive(prim) => Some(prim.to_string()),
             Union(variants) => {
-                let variants: Vec<_> = variants.iter().map(|x| &x.1).collect();
+                let parenthesize =
+                    // never must be parenthesized
+                    variants.is_empty() ||
+                    // named union must be parenthesized
+                    variants.iter().any(|(n, _)| n.is_some());
 
-                SeparatedExprs {
+                let variants: Vec<_> = variants.iter().map(|(n, t)| UnionVariant(n, t)).collect();
+                let sep_exprs = SeparatedExprs {
                     exprs: &variants,
                     inline: " || ",
                     line_end: " ||",
+                };
+
+                if parenthesize {
+                    sep_exprs.write_between("(", ")", opt)
+                } else {
+                    sep_exprs.write(opt)
                 }
-                .write(opt)
             }
             Singleton(lit) => Some(lit.to_string()),
             Tuple(elements) => SeparatedExprs {
@@ -99,5 +109,20 @@ impl WriteSource for TupleField {
                 Some(r)
             }
         }
+    }
+}
+
+struct UnionVariant<'a>(&'a Option<String>, &'a Ty);
+
+impl WriteSource for UnionVariant<'_> {
+    fn write(&self, mut opt: WriteOpt) -> Option<String> {
+        let mut r = String::new();
+        if let Some(name) = &self.0 {
+            r += name;
+            r += " = ";
+        }
+        opt.consume_width(r.len() as u16);
+        r += &self.1.write(opt)?;
+        Some(r)
     }
 }